+++ Project 3/api/photos.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const { validateAgainstSchema, extractValidFields } = require('../lib/validation');
+const { requireAuthentication, authenticate } = require('../lib/auth');
 
 const db = require('../lib/connections');
 
@@ -9,8 +10,8 @@ exports.router = router;
  * Schema describing required/optional fields of a photo object.
  */
 const photoSchema = {
-  userid: { required: true },
-  businessid: { required: true },
+  userId: { required: true },
+  businessId: { required: true },
   caption: { required: false }
 };
 
@@ -57,21 +58,36 @@ async function deletePhotoById(photoid) {
   return result.affectedRows > 0;
 }
 
+async function getUserFromPhoto(photoid) {
+  const [result] = await db.query(
+    'SELECT userId FROM photos WHERE id = ?',
+    [photoid]
+  )
+
+  return result[0].userId;
+}
+
 /*
  * Route to create a new photo.
  */
-router.post('/', async function (req, res, next) {
+router.post('/', requireAuthentication, async function (req, res, next) {
   try {
-    const id = await insertNewPhotos(req.body);
-    const photo = extractValidFields(req.body, photoSchema);
+    if (authenticate(req.body.userId, req)) {
+      const id = await insertNewPhotos(req.body);
+      const photo = extractValidFields(req.body, photoSchema);
 
-    res.status(201).json({
-      id: id,
-      links: {
-        photo: `/photos/${id}`,
-        business: `/businesses/${photo.businessid}`
-      }
-    })
+      res.status(201).json({
+        id: id,
+        links: {
+          photo: `/photos/${id}`,
+          business: `/businesses/${photo.businessId}`
+        }
+      })
+    } else {
+      res.status(403).json({
+        error: "Unable to insert photo into db."
+      });
+    }
   } catch (err) {
     res.status(400).json({
       error: "Request body is not a valid photo object"
@@ -101,7 +117,7 @@ router.get('/:photoID', async function (req, res, next) {
 /*
  * Route to update a photo.
  */
-router.put('/:photoID', async function (req, res, next) {
+router.put('/:photoID', requireAuthentication, async function (req, res, next) {
   const photoid = parseInt(req.params.photoID);
 
   if (validateAgainstSchema(req.body, photoSchema)) {
@@ -109,27 +125,34 @@ router.put('/:photoID', async function (req, res, next) {
       * Make sure the updated photo has the same businessid and userid as
       * the existing photo.
       */
-    let updatedPhoto = extractValidFields(req.body, photoSchema);
-    let existingPhoto = await getPhotoById(photoid);
-    if (updatedPhoto.photoid === existingPhoto.photoid && updatedPhoto.userid === existingPhoto.userid) {
-      try {
-        updateStatus = await updatePhotoById(photoid, req.body);
-        if (updateStatus) {
-          res.status(200).json({
-            links: {
-              photo: `/photos/${photoid}`,
-              business: `/businesses/${updatedPhoto.businessid}`
-            }
-          });
-        } else {
-          next();
+    userId = await getUserFromPhoto(photoid);
+    if (authenticate(userId, req)) {
+      let updatedPhoto = extractValidFields(req.body, photoSchema);
+      let existingPhoto = await getPhotoById(photoid);
+      if (updatedPhoto.photoId === existingPhoto.photoId && updatedPhoto.userId === existingPhoto.userId) {
+        try {
+          updateStatus = await updatePhotoById(photoid, req.body);
+          if (updateStatus) {
+            res.status(200).json({
+              links: {
+                photo: `/photos/${photoid}`,
+                business: `/businesses/${updatedPhoto.businessId}`
+              }
+            });
+          } else {
+            next();
+          }
+        } catch (err) {
+          error: "Unable to update photo."
         }
-      } catch (err) {
-        error: "Unable to update photo."
+      } else {
+        res.status(403).json({
+          error: "Updated photo cannot modify photoid or userid"
+        });
       }
     } else {
       res.status(403).json({
-        error: "Updated photo cannot modify photoid or userid"
+        error: "Unauthorized to perform this action."
       });
     }
   } else {
@@ -142,14 +165,21 @@ router.put('/:photoID', async function (req, res, next) {
 /*
  * Route to delete a photo.
  */
-router.delete('/:photoID', async function (req, res, next) {
+router.delete('/:photoID', requireAuthentication, async function (req, res, next) {
   const photoid = parseInt(req.params.photoID);
   try {
-    deleteStatus = await deletePhotoById(photoid);
-    if (deleteStatus) {
-      res.status(204).end();
+    userId = await getUserFromPhoto(photoid);
+    if (authenticate(userId, req)) {
+      deleteStatus = await deletePhotoById(photoid);
+      if (deleteStatus) {
+        res.status(204).end();
+      } else {
+        next();
+      }
     } else {
-      next();
+      res.status(403).json({
+        error: "Unauthorized to perform this action."
+      });
     }
   } catch (err) {
     res.status(500).send({
